+++ src/components/Header/components/UpperNav.tsx
@@ -7,14 +7,7 @@ import Link from "../../Link/Link";
 import List from "../../List/List";
 import { upperNavLinks } from "../utils/headerUtils";
 
-export interface LoginProps {
-  catalogRef?: React.RefObject<HTMLAnchorElement>;
-  greetingRef?: React.RefObject<HTMLDivElement>;
-  isMobile?: boolean;
-  patronName: string;
-}
-
-const UpperNav = chakra(({ patronName }: LoginProps) => {
+const UpperNav = chakra(() => {
   const styles = useMultiStyleConfig("HeaderUpperNav", {});
 
   return (
@@ -23,7 +16,7 @@ const UpperNav = chakra(({ patronName }: LoginProps) => {
         id="header-nav-upper"
         inline
         listItems={[
-          <HeaderComponents.LoginButton key="login" patronName={patronName} />,
+          <HeaderComponents.LoginButton key="login" />,
           <Link href={upperNavLinks.locations} key="locationsLink">
             Locations
           </Link>,