+++ src/utils/componentCategories.ts
@@ -78,6 +78,7 @@ const categories = {
     title: "Hooks",
     components: [
       "useCarouselStyles",
+      "useCloseDropDown",
       "useNYPLBreakpoints",
       "useNYPLTheme",
       "useWindowSize",